requirements.txt
pyserial~=3.5
nidaqmx>=1.0.1
pandas>=2.0.3
simple_colors~=0.1.5
opencv-python~=4.10.0.84
gtts~=2.5.3
SpeechRecognition==3.10.4
WMI~=1.5.1
# auto-py-to-exe~=2.44.1
pyinstaller==6.11.0
# serial>=0.0.97
openpyxl>=3.1.5
# numpy==1.24.4
# matplotlib==3.7.5
pillow==10.4.0
ImageHash>=4.3.1
memory_profiler~=0.61.0